+++ smart-contracts/test/Agora.t.sol
@@ -1,14 +1,14 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.24;
 
-import 'forge-std/Test.sol';
+import "forge-std/Test.sol";
 
-import { IAgora } from '../contracts/polity/interfaces/IAgora.sol';
-import { ICitizen } from '../contracts/polity/interfaces/ICitizen.sol';
+import {IAgora} from "../src/polity/interfaces/IAgora.sol";
+import {ICitizen} from "../src/polity/interfaces/ICitizen.sol";
 
-import { Agora } from '../contracts/polity/Agora.sol';
-import { CitizenRegistry } from '../contracts/polity/CitizenRegistry.sol';
-import { Vote } from '../contracts/polity/Vote.sol';
+import {Agora} from "../src/polity/Agora.sol";
+import {CitizenRegistry} from "../src/polity/CitizenRegistry.sol";
+import {Vote} from "../src/polity/Vote.sol";
 
 contract ProposeTest is Test {
     Agora agora;
@@ -37,7 +37,7 @@ contract ProposeTest is Test {
     }
 
     function testCreateIPFS() public {
-        agora.createIPFS(A, 'bafkreigykb62xhd7gluyfzdv2opzgkbgovtphi2fuyjpdygbilp6rdchsu');
+        agora.createIPFS(A, "bafkreigykb62xhd7gluyfzdv2opzgkbgovtphi2fuyjpdygbilp6rdchsu");
 
         Agora.Proposal[] memory page = agora.proposals(0, 100);
 
@@ -45,17 +45,17 @@ contract ProposeTest is Test {
     }
 
     function testVotesThresholdOf() public {
-        uint[] memory values = new uint[](1);
+        uint256[] memory values = new uint256[](1);
         values[0] = 0;
         address[] memory targets = new address[](1);
         targets[0] = address(this);
         bytes[] memory calldatas = new bytes[](1);
-        calldatas[0] = abi.encodeCall(ICitizen.recordApprovedEvent, (address(0x1234), 'DUMMY_CID'));
+        calldatas[0] = abi.encodeCall(ICitizen.recordApprovedEvent, (address(0x1234), "DUMMY_CID"));
 
-        uint256 id = agora.propose(targets, values, calldatas, 'test');
+        uint256 id = agora.propose(targets, values, calldatas, "test");
         uint256 snap = agora.proposalSnapshot(id);
 
-        vm.expectRevert(bytes('Voting not started'));
+        vm.expectRevert(bytes("Voting not started"));
         agora.votesThresholdOf(id);
 
         vm.roll(snap + 1); // vm.roll(n) instantly sets the next block number to n in Forge tests.
